requirements.txt
asgiref==3.3.1
astroid==2.5.1
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-gis==0.17
geographiclib==1.50
geopy==2.1.0
isort==5.7.0
lazy-object-proxy==1.5.2
leaflet==0.0.3
mccabe==0.6.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.7.2
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
